app/requirements.txt
Django>=3.2,<4.0
django-configurations
djangorestframework
gunicorn
psycopg2-binary
python-decouple
matplotlib
reportlab
scikit-learn
joblib
pandas
requests